# [foundation] Cache parts of `NSObject.ConformsToProtocol`