--- a/src/Foundation/NSObject2.cs
+++ b/src/Foundation/NSObject2.cs
@@ -477,9 +477,9 @@ namespace Foundation {
 
 			// the linker/trimmer will remove the following code if the dynamic registrar is removed from the app
 			var classHandle = ClassHandle;
-			lock (protocol_cache) {
+			lock (Runtime.protocol_cache) {
 #if NET
-				ref var map = ref CollectionsMarshal.GetValueRefOrAddDefault (protocol_cache, classHandle, out var exists);
+				ref var map = ref CollectionsMarshal.GetValueRefOrAddDefault (Runtime.protocol_cache, classHandle, out var exists);
 				if (!exists)
 					map = new ();
 				ref var result = ref CollectionsMarshal.GetValueRefOrAddDefault (map, protocol, out exists);
@@ -487,10 +487,10 @@ namespace Foundation {
 					result = DynamicConformsToProtocol (protocol);
 #else
 				bool new_map = false;
-				if (!protocol_cache.TryGetValue (classHandle, out var map)) {
+				if (!Runtime.protocol_cache.TryGetValue (classHandle, out var map)) {
 					map = new ();
 					new_map = true;
-					protocol_cache.Add (classHandle, map);
+					Runtime.protocol_cache.Add (classHandle, map);
 				}
 				if (new_map || !map.TryGetValue (protocol, out var result)) {
 					result = DynamicConformsToProtocol (protocol);
@@ -500,8 +500,6 @@ namespace Foundation {
 				return result;
 			}
 		}
-
-		static Dictionary<NativeHandle, Dictionary<NativeHandle, bool>> protocol_cache = new ();
 
 		bool DynamicConformsToProtocol (NativeHandle protocol)
 		{