@vs-mobiletools-engineering-service2 2022-06-17
## :white_check_mark: API diff for current PR / commit

<details><summary>Legacy Xamarin (No breaking changes)</summary>

* ~iOS~ (no change detected)
* ~tvOS~ (no change detected)
* ~watchOS~ (no change detected)
* ~macOS~ (no change detected)

</details>

<details><summary>NET (empty diffs)</summary>

* ~iOS~: (empty diff detected)
* ~tvOS~: (empty diff detected)
* ~MacCatalyst~: (empty diff detected)
* ~macOS~: (empty diff detected)

</details>


## :white_check_mark: API diff vs stable

<details><summary>Legacy Xamarin (No breaking changes)</summary>

* iOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/ios-api-diff.html) [gist](https://gist.github.com/bcf2e3e386b1cc5c3d9e44e6ad626f13) (No breaking changes)
* tvOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/tvos-api-diff.html) [gist](https://gist.github.com/173c13fad780e2fda223c200776c637e) (No breaking changes)
* watchOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/watchos-api-diff.html) [gist](https://gist.github.com/56f6e9fab6932d5d330e9ffeccd425ed) (No breaking changes)
* macOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/mac-api-diff.html) [gist](https://gist.github.com/b40e4a21dc8a95987cf170c1632deead) (No breaking changes)

</details>

<details><summary>.NET (No breaking changes)</summary>

* iOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/diff/dotnet/Microsoft.iOS.Ref/ref/net6.0/Microsoft.iOS.html) [gist](https://gist.github.com/ce0f1a8c11e4df46565992ce21c43112) (No breaking changes)
* tvOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/diff/dotnet/Microsoft.tvOS.Ref/ref/net6.0/Microsoft.tvOS.html) [gist](https://gist.github.com/5e94b548ed8b195ad13c8bff8e577cfb) (No breaking changes)
* MacCatalyst: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/diff/dotnet/Microsoft.MacCatalyst.Ref/ref/net6.0/Microsoft.MacCatalyst.html) [gist](https://gist.github.com/76d7a8e06448d43fa038ea71c7b19c6c) (No breaking changes)
* macOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/diff/dotnet/Microsoft.macOS.Ref/ref/net6.0/Microsoft.macOS.html) [gist](https://gist.github.com/f4a9a10b47ae87dbc160a01638658b11) (No breaking changes)
* Microsoft.iOS vs Microsoft.MacCatalyst: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/diff/dotnet/iOS-MacCatalyst-diff/Microsoft.iOS.Ref/ref/net6.0/Microsoft.iOS.MacCatalyst.html) [gist](https://gist.github.com/4a70e56f0042003696abd055c70cdf9b)

</details>

<details><summary>Legacy Xamarin (stable) vs .NET</summary>

* iOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/diff/dotnet/legacy-diff/Microsoft.iOS.Ref/ref/net6.0/Microsoft.iOS.html) [gist](https://gist.github.com/ae24928a1be72db44baecbaffee1ff6b)
* tvOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/diff/dotnet/legacy-diff/Microsoft.tvOS.Ref/ref/net6.0/Microsoft.tvOS.html) [gist](https://gist.github.com/3f1c04dd62a36c6607f44eb9b972636d)
* macOS: [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/detected-changes/20220617.7/6302386/;/stable-api-comparison/diff/dotnet/legacy-diff/Microsoft.macOS.Ref/ref/net6.0/Microsoft.macOS.html) [gist](https://gist.github.com/17b6d0d3ada128195e90d36ffcaa28e9)

</details>


## :white_check_mark: Generator diff

Generator diff is empty


[Pipeline](https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=6302386&view=ms.vss-test-web.test-result-details) on Agent 
Hash: [ddf76f4b8daee8fedc72049f63278174aa482fe2](https://github.com/xamarin/xamarin-macios/pull/15294/commits/ddf76f4b8daee8fedc72049f63278174aa482fe2)  [PR build]