@vs-mobiletools-engineering-service2 2022-06-17
# :computer: [PR Build] Tests on macOS M1 - Mac Big Sur (11.5) passed :computer:

:white_check_mark: **All** tests on macOS M1 - Mac Big Sur (11.5) passed.

[Pipeline](https://dev.azure.com/devdiv/DevDiv/_build/index?buildId=6302386&view=ms.vss-test-web.test-result-details) on Agent 
Hash: [ddf76f4b8daee8fedc72049f63278174aa482fe2](https://github.com/xamarin/xamarin-macios/pull/15294/commits/ddf76f4b8daee8fedc72049f63278174aa482fe2)  [PR build]