@vs-mobiletools-engineering-service2 2022-06-18
### :x: [CI Build] Tests failed on VSTS: simulator tests iOS :x:

Tests failed on VSTS: simulator tests iOS. 
* [Azure DevOps](https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=6302386&view=ms.vss-test-web.test-result-details)
* [Html Report (VSDrops)](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20220617.7/6302386/simulator/;/tests/vsdrops_index.html) [Download](https://devdiv.visualstudio.com/DevDiv/_apis/build/builds/6302386/artifacts?artifactName=HtmlReport-simulator&api-version=6.0&$format=zip)


# Test results

<details>
<summary>1 tests failed, 147 tests passed.</summary>

## Failed tests

 * mono-native-compat/watchOS 32-bits - simulator/Debug: Crashed
</details>


[Pipeline](https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=6302386&view=ms.vss-test-web.test-result-details) on Agent XAMBOT-1105.Monterey'
Merge ddf76f4b8daee8fedc72049f63278174aa482fe2 into 42c1c66a142e4f2deed095f3f91fa2da4dfa6d12