@vs-mobiletools-engineering-service2 2022-06-21
# :x: [PR Build] Tests on macOS Mac Catalina (10.15) failed :x:

Tests on macOS Mac Catalina (10.15) failed for unknown reasons.

[Pipeline](https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=6311035&view=ms.vss-test-web.test-result-details) on Agent 
Hash: [d03f9362806fe3c75fe7b54508a2259df6522e6b](https://github.com/xamarin/xamarin-macios/pull/15294/commits/d03f9362806fe3c75fe7b54508a2259df6522e6b)  [PR build]